[crypto] Rearrange P-256 code to save code size.
ECDH was the only call site for `p256_scalar_mult`, so this commit
renames it to `p256_shared_key` and further specializes it to ECDH
shared-key generation.  Then, it can be moved to `p256_a2b.s`, which I
then renamed to `p256_shared_key.s` to better fit its new purpose.

Signed-off-by: Jade Philipoom <jadep@google.com>