[keymgr_dpe, dv] updates error handling checks in scb
In keymgr_dpe scb updates the conditions for signaling
an error for invalid_op, sw_invalid_input and hw_invalid_input.

For get_invalid_op, the conditions are related to if the operation
adheres to using a valid src key, within the limits of boot_stage,
key version and key_policy.

For get_sw_invalid_input it's based on key_version and a valid src slot.

For get_hw_invalid_input no longer valid states are removed and the checks
are not general for the available working state.

In process_update_after_op_done() removed the fork/begin statement that
stalls a neg posedge cc beore updating the expected current_op_status.
This method is no longer needed. The expected status, can be OpWip, when
SW reads the op_status csr. By doing the fork and stall, the internal key
is already updated, by the time the error checks are done. These error checks
will end up being done with the new key data rather than the old which will result
in false fails.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>