[edn] Reset auto cmd FIFOs when EDN is disabled
This commit resets the auto mode cmd FIFOs whenever the EDN is disabled.
This is done by setting the main_sm_done_pulse_o to high whenever the
edn_enable_i signal changes to low. In edn_core.sv, the done_pulse is used
for several purposes other than just resetting the FIFOs. However, this change
has no or little impact for the other three use cases of this signal. In two of
those three cases the done_pulse has no impact, since the EDN is disabled and the
multiplexer ignores the pulse. In the third case the max_num_between_reseeds is
loaded into a counter whenever the done_pulse is high. This should be fine since the
CSRNG is initialized anyways, should the EDN enter auto mode at a later time.
The only difference for the third case would be that when the auto mode is left
through disabling the EDN, the counter is now reset and should it re-enter auto
mode the counter starts counting from 0 again. This didn't use to be the case.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>