[ci] Address @bitstreams//BUILD.bazel nullifying action cache
Place bitstreams from other jobs in a consistent location, so the
BUILD.bazel file and the command line for the action are not needlessly
changed for every CI run.

Originally, the git revision would be part of the path name to source
files, and this precluded any possibility of getting a hit to the action
cache.

Signed-off-by: Alexander Williams <awill@opentitan.org>