[otbn,doc] Extract list of ISRs into YAML files
No functional change, except that the lists of CSRs/WSRs in the
documentation are now auto-generated from a big YAML file.

The plan is to use that YAML file for the other lists that you see
mentioned in e.g. otbn/README.md. But I grabbed the initial list from
the documentation, so "generating itself" seemed like the obvious
first step.

This should be the first step to sorting out issue #3323, which tracks
whether the toolchain can support named CSRs and WSRs in OTBN assembly
code.

If reviewing this PR:

 - You can mostly ignore the .yml files: the contents get used to
   generate README.md, which is also checked in.

 - The script in isr.py is the thing that does the easy job of parsing
   the yaml.

 - The script in md_isrs.py is the thing that converts the parsed yaml
   into a markdown table.

 - README.md is now auto-generated, and is intended to be mostly
   unchanged.

Note that I haven't quite managed to auto-generate the same README.md
as was there before. The changes are:

 - Using *bold* instead of <strong>bold</strong>

 - Avoiding some stray capitals in Flag Group

 - Indenting everything more uniformly. For example, see the RND WSR.

 - Adding anchor tags to all WSRs. We needed this for WSRs like
   KEY_S0_L, and I thought it seemed cleaner to do it uniformly
   instead of adding a flag to the YML to determine whether the WSR
   needed an anchor tag.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>