[prim_sha2,rtl] Add RTL implementation + update core + lint waivers
This adds the RTL implementation for the merged SHA-2 prim that can be
configured at compile-time as either multi-mode or 256-bit.
The RTL includes the 32-bit word input wrapper prim_sha2_32 that
instantiates the prim_sha2. This also fixes many coding style issues
that were in the original HMAC IP RTL, and updates the relevant
prim core files. This also updates the lint waivers and RTL fixes,
to pass AscentLint linting.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>