[dv/dpi] improvements to tcp_server receive loop
The server thread now breaks out of the receive loop if there's no
room in the input buffer instead of spinning inside
tcp_buffer_put_byte(). This avoids a deadlock where the jtagdpi thread
is blocked on a full output buffer, but the server thread is spinning
waiting for the jtagdpi thread to make room in the input buffer.

It also treats a failure to receive any data (after a positive
FD_ISSET on the read_fds) as a client disconnect.

Signed-off-by: Kip Walker <kip@rivosinc.com>