[ci] Make hyperdebug tests blocking
This `continueOnError` setting allows other jobs to start even if this
one fails. The hyperdebug tests happen at the end of the pipeline
anyway, so this doesn't help us very much.

The flag does prevent us from using the "re-run failed jobs" feature of
Azure, though. Azure doesn't seem to think the job failed, so we have to
force-push to the PR to get it to re-trigger the whole CI pipeline.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>