[opentitantool] Properly support flashing by UART
Most flashing protocols relies on shared code in `bootstrap/mod.rs` to
apply the `ROM_BOOTSTRAP` strapping and reset the device to enter
bootstrapping mode.

However, the serial protocol to be implemented in ROM_EXT as well
as another serial "rescue" protocol used by legacy Google Titan chips
uses only the RESET pin and the UART RX/TX signals.  As such, they need
to sidestep the common logic, which they request by returning false from
`uses_common_bootstrap_reset()`.

It seems however, that a call to `transport.reset_target()` in
`bootstrap/mod.rs` has been inserted, without being made conditional on
`uses_common_bootstrap_reset()`, and in such a way as to prevent
the `leave_in_reset` flag from working.  (That is, reset would always be
asserted then released as the final operation of a bootstrap, even if
the called asked for reset to be asserted, and kept asserted until the
caller was ready to let the newly flashed image boot for the first
time.)

This CL restores the functionality of `leave_in_reset`.

Change-Id: I6b7cf5f6a11d33a5aba84814abc2f061248e0914
Signed-off-by: Jes B. Klinke <jbk@chromium.org>