[manuf] add device ID parameter to HwCfg individualize function
During individualization, the device ID is copied from flash info page 0
to the HwCfg OTP partition. However, the device ID in flash is not
checked, and blindly copying it makes it difficult to test both CP and
FT stages individually, since to test FT provisioning would require
setting up flash info pages before test execution. Moreover, if the
device ID is not checked for all 0s, it will cause the keymgr to lock up
later during the provisioning flow.

This adds a device ID parameter to the HwCfg individualization function
to override the device ID stored in flash (if it is all 0s, i.e., it is
missing), or to check against what is already in flash to ensure the
provisioning payload is the one expected.

Signed-off-by: Tim Trippel <ttrippel@google.com>