[sca] Switch AES SCA to uJson
For the penetration test at the silicon, we need to switch from
the current simpleserial protocol to the uJSON protocol for the
AES SCA tests.

This PR implements the device side command handler. To be
compatible to existing tests, the previous simpleserial commands
were copied and adapted to the new protocol.

For now, we want to keep the code in
sw/device/tests/crypto/cryptotest, we might move it to the
sca directory in the future.

The host code is located in the ot-sca repository.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>