[spi_passthru] Support READ_4B and FAST_READ_4B
The default SPI eeprom on the CW310 board is an ISSI25WP016 and it does
not support any 4-byte commands.  We do, however, want to support the
explict 4-byte opcodes in the passthru implementation and they can be
tested in an ad-hoc fashion by connecting a different eeprom to the
CW310 board via `USR_SPI0` header `P1`.

1. Fix an error in the spi_passthru_test definition: explicitly load a
   bitstream so that we make sure we're testing with the expected
   configuration.
2. Add READ_4B and FAST_READ_4B to the spi device configuration.
3. Support explicit 4-byte opcodes in opentitanlib/tool.  I don't
   believe the SFDP does a good job of encoding wither or not these
   opcodes are supported, so they need to be enabled via a command line
   flag.

Signed-off-by: Chris Frantz <cfrantz@google.com>