[manuf] add third personalization binary to gen attestation keys
A third (and final) personalization binary is added to advance the
keymgr and generate attestation keys / certificates. This code was
placed in a separate binary as a short term fix to name collisions
between functions in the silicon_creator OTBN driver (used by the
`otbn_boot_services` lib that generates the attestation keys), and the
cryptolib OTBN driver (used in the `personalization` lib that provisions
OTB Secret2 partition and the keymgr flash info pages).

Signed-off-by: Tim Trippel <ttrippel@google.com>