[adc_ctrl,dv] Tidy up access to intr_state in env_cfg files
This code was duplicated in lots of places: I think probably from a
template file in the first place. In each case, if the code compiles,
then ${MY_BLOCK}_reg_block must have a field called intr_state. In
that situation get_reg_by_name() will always return something that is
not null, so the test in the if statement will always be true.

Use the shorter, more direct, form each time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>