[csrng/dv] Add coverage for the transitions of the fips bit
This commit adds the coverage for the transitions of the fips flag/bit
at the command input and genbits output.
The transitions for the commands are recorded at each INS or RES command,
since only those can cause a transition in the fips bit at the output. For
the commands, the coverage now also records whether the INS/RES commands
result in a new instantiation or not.
For the fips flag at the register output, the valid bit is always recorded,
while the fips bit is only recorded when the valid bit is high.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>