[util] Ignore Verilator stderr in version check
With funny CI environments, the locale might not be set properly,
making perl complain. If we capture the warning message as well as the
version number, we'll end up failing the version check.

Silently drop the stderr here: all we need to know is whether
Verilator responds with a sensible version.

Co-authored-by: Stefan Wallentowitz <stefan.wallentowitz@hm.edu>
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>