[dv/jtag] Add capability for cycles in run-test-idle state
Add a capability to sequence items to dwell in the run-test-idle state
for a configurable number of JTAG cycles. This is critical for
interacting with RISC-V DTMs, so they may have enough cycles to complete
DMI transactions before the TAP reaches the Capture state. If there are
insufficient JTAG cycles, the DTM will error out, and no forward
progress will be made.

Signed-off-by: Alexander Williams <awill@opentitan.org>