[silicon_creator/keymgr] add function to set Owner max key version
This adds a function to the silicon_creator keymgr driver to set the max
key version. Additionally this updates the on-host unit tests and
on-device functest to test this additional function.

Lastly, this updates the on-device functest to crank the keymgr through
all possible key states, as will be done for both ES and PROD chips with
the updated attestation flow.

Signed-off-by: Tim Trippel <ttrippel@google.com>