[dif] Add keymgr_dpe initialize function.
The first keymgr_dpe advance call is used to initialize the key manager
and latch the UDS key in one of the destination slots. Software binding,
max version and policy registers are not used by hardware during this
operation.

This change also updates the `keymgr_dpe_testutils.c` to use the new
initialization function.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>