[sival] SiVal Bring-Up SKU
This commit explicitly adds all the OTP partitions to the
earlgrey_a0_skus/siva_bringup/BUILD file. The CREATOR_SW and OWNER_SW
partitions are updated to explicitly include all OTP fields supported by
the OTP map. If the field was previously not set, it is initialized with
a value equivalent to 0. The OTP configuration tool will be updated to
skip writes of OTP zero values.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>