[util] Create modes as a dictionary keyed by name
This commit is a little silly: it changes the Modes.create_modes()
static function to return a dictionary, then changes all the call
sites to cast the result back to a list of the values in the
dictionary.

The point is that we will be able to do a follow-up that removes some
of this casting, but dividing things up like this should make the
incremental steps a bit simpler.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>