[dma/dv] Sample `error_code_cg` when reading `ERROR_CODE` register
Prior to this commit, `error_code_cg` was sampled when the `STATUS`
register was read, which would provide incorrect values.  (It may have
been correct at some point when error code bits were part of the status
register.)

One complication comes from the check that needs to know if bus errors
have happened when the `STATUS` register is read: prior to this commit,
this check used the `bus_error` bit of the `ERROR_CODE` register (or
rather it was supposed to use that bit but did not because of the
incorrect value), which is not directly available when the `STATUS`
register is read.  This check now instead uses
`src_tl_error_detected | dst_tl_error_detected`, which are scoreboard
variables that track errors based on TL-UL responses.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>