[util] Treat build_modes as a dictionary, not a list, in dvsim
Before this change, the build_modes field in an hjson file was a list
of objects. We looked up build mode by searching through that list for
an object with the expected name field.

This change switches things so that it is genuinely a dictionary,
keyed by build mode name. This shouldn't make much difference, except
that merging lists becomes possible. Before this change, you might
have a situation where you wanted to merge

   [{name: foo}, {name: bar}]

with

   [{name: foo}, {name: baz}].

The result would be

   [{name: foo}, {name: bar}, {name: foo}, {name: baz}]

which is a bit sad!

With this change, you merge:

   {foo: {}, bar: {}}

and

   {foo: {}, baz: {}}

and get

   {foo: {}, bar: {}, baz: {}}

Of course, the objects are a bit more interesting (not just {}!), but
the list/dict switch should be right.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>