[silicon_creator] Reformat error enum to make changes easier
With the current layout, any change ends up modifying the last line
of the enumeration which is not ideal for atomicity of changes.
This commit reformats the enum slightly so we can add changes on new
lines every time. This requires to add a trailing comma to the enum
and therefore requires a couples of fixes in the code.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>