[tlul/rtl] Change `ASSERT_I` to `ASSERT` to fix erroneous assertion failure
`ASSERT_I` is not disabled during reset or when reset is unknown, so it
fails erroneously.  This caused a large number of tests (32 according to
the current report) in Darjeelings `xbar_main` regressions to fail.

This commit fixes that problem by changing the assertion to `ASSERT`,
which is disabled when reset is not deasserted (thus it's also disabled
when reset is unknown).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>