[otbn,dv] Wait a bit longer for secure wipe at start of vseq
This was sometimes failing. It seems that the "wait 500 cycles" delay
that was in the code was on the edge, and got pushed over if the EDN
clk_rst_if happened to stay in reset for a bit too long.

Explicitly wait to be out of reset and then divide the wait blocks
between the two clocks in a way that should hopefully be a bit more
robust.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>