[ot_certs/ottool] Add command to codegen a certificate template
This command takes a template and outputs a header and source file
that exposes a clean API to generate a certificate. This API
takes as input a structure containing the values of the variables.
The API exposes two functions: one to generate the TBS and one to
generate the certificate, this way it can be used for signing
as well.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>