[usbdev] Predictor of usbdev/[usbdev]usb20_monitor
This request contains the usbdev_TransactionManager class of the usbdev predictor (reference model).In this class the whole flow for the usb predictor is done. It also consist of a modified packet classifier class in which the pid identifier values are changed according to the new usb20_item.

This request also contains usb20_monitor with SOF packet

Signed-off-by: sarahmazhar-impare <sarah.impare@gmail.com>