[ot_certs] Add DER implementation of the ASN1 builder trait
This implementation can produce the DER representation of an ASN1
document. It only supports literal values and will error out if
a variable is provided. Unittest are provided to make sure that
the output is correct. Most of the tests comes from, or are
variations of, examples in the X.690 specification.

This implementation serves two purposes: illustrate how to
implement the builder trait and provide a rust refernce to compare
against. This will be useful when testing the upcoming codegen
implementation.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>