[opentitantool] Clean up OpenOCD arguments to opentitantool
rules/opentitan/openocd.bzl unnecessarily set --openocd-adapter-config
even when using HyperDebug backend.  This used to be just ignored, but
the next CL will change the behavior of that flag, such that specifying
it for a HyperDebug backend means that the user does not want to use the
built-in JTAG functionality of HyperDebug, but wants to use an external
JTAG adapter given in the flag.

Also, in a single place `opentitan_functest()` was used to test JTAG
functionality of the HyperDebug/CW310 setup.  This is now more cleanly
tested using `opentitan_test()` and one of the SiVal environments.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: I6ea4c2818d9083e296477a40944d574b313b413d