[manuf] Fix to broken test
Previous PRs #20619 and #20627 interfered in a way to break the
cp_provision test, as the former made --openocd-adapter-cfg function for
hyper310, and the latter changed a manufacturing test to run on a
hyper310 execution environment and didn't change the OpenOCD adapter
configuration.

This CL makes the test use `hyper310_jtag_params`, matching the fact
that it uses HyperDebug for interfacing with CW310 (or the ASIC).

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I8f4179b42dd7ce4675cf7212f4b3964622b70b78