[otbn,dv] Fix start_running_otbn in otbn_base_vseq
This task is supposed to start an OTBN run and then finish when the
run is still in progress.

A recent change to this vseq altered things so that "in progress"
includes "in the middle of the secure wipe at the end". Unfortunately,
this change can break some vseqs. In particular,
otbn_rf_base_intg_err_vseq expects to be able to corrupt the RF and
have it cause an error. I think an opportune time to do so is no
longer guaranteed to happen after the run completes and we're in the
middle of secure wipe.

Tighten the definition again, which I think will fix the problem.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>