[es/dv] Get intr vseq working again
The old vseq fails a lot. This commit removes the err task of the intr vseq
and the test_es_health_test_failed task. The commit also adapts the other
tasks such that no errors happen anymore. The err task is already part
of the err vseq and a reworked version of the test_es_health_test_failed vseq
will be added in the next commit.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>