[es/dv] Add test_es_health_test_failed task to intr vseq
This commit adds the new test_es_health_test_failed task to the intr vseq.
The new version not only checks for the interrupts if the health test fails,
it also checks whether the alert state is entered and no further entropy can
be consumed. Now the task also tests the health test failures in all of the
states that can transition into the alert state.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>