[dv,pwrmgr] Fix pwrmgr_escalation_timeout test
The number of cycles esc_clk need to stop to trigger a timeout is unpredictable:
- The timeout counter can start at any number from 0 to 7.
- The reqack synchronizer adds some variable number of cycles due to its phase.

Avoid spurious failures by reducing the number of cycles of stoppage when
not expecting alerts, and wait reduce the number of cycles to trigger an expected
alert.

Also, avoid running the case that fails due to issue #20516.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>