[sca] Prepare KMAC for discrete
Currently, the SCA code for KMAC uses features to improve SCA
that are only available on the FPGA but not on discrete.

With this PR, the user can transmit a fpga_mode flag (c.f. #269)
allowing to specify which KMAC SCA version to use. For
fpga_mode=1, the existing code is  used.
For fpga_mode=0 (i.e., discret chip), the SCA specific
`sha3_ujson_absorb` function is replaced by calls to DIF:
- dif_kmac_mode_kmac_start
- dif_kmac_absorb
Moreover, the SCA specific `kmac_get_digest` function is
replaced by:
- dif_kmac_squeeze

By using the KMAC DIF functions, we now comply to the intended
usage of KMAC.

Closes #20674

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>