[ot_certs] Replace hjson Value<T> parser by the subst module
The current code is not ideal because it merges both hjson parsing
(i.e. string/array/integer) and conversion to whatever type a
certificate field might have. This requires some special hackery
in the form of the hjson module and leads to poor error messages
and a difficulty to understand what happens.

This commit changes the hjson parser to instead use the `subst`
module to separate concerns: the hjson is first parsed as a
`SubstValue` (ie just a raw string/integer/array) and then is
converted to whatever type the field has using the `ConvertValue`
trait. This means that all the conversion logic is now contained
in `subst` and independent of parsing.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>