[spi_device] Add pipelined reads for flash / passthrough
Add a 2-stage pipeline to break up the combinatorial path for
passthrough reads, particularly for quad output reads.

For selected commands in the CMD_INFO tables, this pipeline samples the
read data from the downstream SPI flash and latches it in spi_device,
then shifts it out to the host 2 cycles later. To use it, the user would
read the number of dummy cycles required for the downstream SPI flash,
configure the CMD_INFO table for the same number as the downstream SPI
flash, then add 2 more cycles and advertise the higher value in an
intercepted SFDP read.

On quad output reads, the host thinks there are N+2 dummy cycles (which
it learned from the SFDP tables advertised by spi_device), and the
downstream SPI device will shift out one additional byte (which will be
harmlessly dropped).

Signed-off-by: Alexander Williams <awill@opentitan.org>