[ot_certs] Improve asn1 library tag's handling
The code currently always allocates 3 bytes for the length and then
moves the data forward if we need less bytes than requires. This works
but means that the code could throw a kErrorAsn1BufferExhausted when
producing a document even though the final encoding would actually fit
into the buffer.

This commit changes the behaviour so that the code now allocates only
1 byte for the length initially and moves the data backwards on finish
if necessary. This guarantees that if the final document fits in the
buffer, then the code will successfully produces it. This requires a bit
of shuffling around in the code. It also requires a bit more care to
avoid an accidental buffer overflow when moving the data. There is a
unittest to specifically test corner cases in the moving code to make
sure we never write past the end of the buffer.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>