[ot_certs] Add support for boolean variables
This commit adds support for boolean variables in the template,
ASN1 DER and codegen. This is relatively extensive change because
the flags that were previously just booleans need to be migrated
to use Value<bool>.
This commit also contains the unittests for the parsers and der
generator. The ASN1 codegen uses the `asn1_bitstring_t` facilities
introduced in the previous commit.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>