[otbn] Slightly weaken some loop stack assertions
The existing assertions got triggered if a loop body ended with
another loop like this:

    ...
    loopi 1, 1
    loopi 1, 123
    ...

Here, the outer loop contains the second loopi as its only
instruction. When that instruction runs, we're completing the only
iteration of the outer loop (and popping from the loop stack) at the
same time as we start an inner loop (and pushing to the loop stack).

This isn't allowed!

The hardware correctly kills the program and signals a SW error. This
commit weakens the assertions so that they allow a simultaneous push
and pop, so long as it causes an error.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>