[sival,tests] Fix pwrmgr_random_sleep_all_wake_ups
This test expects sensor_ctrl to have a recoverable error at bit 0,
but the AST triggers an additional alert for bit 5. We clear this
additional bit and this test is working, but we added a TODO
against #20798.
For each wakeup add checks that the CSR in the unit issuing the wakeup
is also signalling a wakeup.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>