[edn/dv] Fix failing edn_err_vseq tests
The edn_err_vseq test seems to be failing for multiple reasons.
Mainly because of a hanging wr_cmd(). This test sequence used
to always write the same couple of commands no matter in what state
the error state is induced. This commit reworks the test sequence
such that only those commands are issued that are really needed to
get the EDN into the wanted ack/main SM state. This way we can avoid
hangs because of unnecessary commands.
Furthermore, the err sequence now waits for a random ack/main SM state
for all types of errors. This used to be the case only for ack_sm, main_sm
and counter errors.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>