[top_darjeeling,dv] Fix `rel_path` of top-level IPs
The path used to publish the reports of top-level IPs for Darjeeling is
currently wrong, as can be seen in the links of the [summary page][1]:
- ALERT_HANDLER:
  https://reports.opentitan.org/hw/top_darjeeling/ip_autogen/alert_handler/dv/latest/report.html
- CLKMGR:
  https://reports.opentitan.org/hw/ip/clkmgr/dv/latest/report.html
- PWRMGR:
  https://reports.opentitan.org/hw/ip/pwrmgr/dv/latest/report.html
- RSTMGR:
  https://reports.opentitan.org/hw/ip/rstmgr/dv/latest/report.html

All links miss the `integrated/` prefix after the domain name, so the
link for `alert_handler` cannot be found and the links for `clkmgr` and
`rstmgr` collide with the results for Earlgrey (`pwrmgr` does not
collide because it lives under `hw/top_earlgrey/ip_autogen/pwrmgr` on
`master`).

When printing the commands used to publish the results (invoking `dvsim`
with `--verbose=debug`), we can see that the absolute path to the local
source path is included in the remote path; for example:
https://reports.opentitan.org/integrated//home/dev/src/hw/ip/clkmgr/dv//latest/report.html
when `dvsim` is run inside an OpenTitan container (the local source path
for the regression runner that publishes the results is likely
different).

The reason for this is that `rel_path` for the top-level IPs in
`hw/top_darjeeing/dv/top_darjeeling_sim_cfgs.hjson` includes
`{proj_root}`, which resolves to an absolute path.

This commit fixes the problem by removing `{proj_root}` (and the
trailing slash) from `rel_path` of Darjeeling's top-level IPs.

[1]: https://reports.opentitan.org/integrated/hw/top_darjeeling/dv/summary/latest/report.html

Signed-off-by: Andreas Kurth <adk@lowrisc.org>