[hyper310] Reduce drive strength for USERIOA and USERIOB
Reduce drive strength of IOBs for the USERIOA and USERIOB connectors to
8 mA. The signal quality has not been measured directly, but it seems
that there is strong enough coupling over the ribbon cables where
simultaneous switching of SPI device I/Os can cause the CSB to be
sampled high in the middle of a transaction, leading to spurious
failures.

Bringing the drive strength down to 8 mA allowed QSPI passthrough tests
to pass on my setup (up to at least 6 MHz). Further experiments may be
needed to find the optimal values.

This changes only the CW310 + hyperdebug image.

Signed-off-by: Alexander Williams <awill@opentitan.org>